Parameterize 'shell.bash'.
Rename shell to shell.bash to make it clear that it is written
for bash (because here/test depends on bats) and might not work
with sh.

Name the files to be optionally sourced '*.src' which is both
(i) phonetically descriptive and (ii) similar to '*.rc' which
is also sourced.

Invoking [shell.bash] -- i.e. without parameters -- sources
'default.src' by default.

Invoking [shell.bash test] will test the system, useful for
git pre-commit hooks.